LayoutTests/http/tests/security/img-with-failed-cors-check-fails-to-load-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: Cross-origin image load denied by Cross-Origin Resource Sharing policy.
ALERT: PASS: The error event was called.
This test passes if the image below does not load. 
